+++ app/src/index.css
@@ -451,3 +451,47 @@ body.dark .pricing-highlights {
 .react-contexify .react-contexify__submenu {
   @apply max-h-[300px] overflow-y-auto;
 }
+
+.convert-on-paste-overlay {
+  background: linear-gradient(
+    to bottom,
+    hsla(251.35, 60.66%, 88.04%, 0) 0%,
+    hsla(251.35, 60.66%, 88.04%, 0.013) 8.1%,
+    hsla(251.35, 60.66%, 88.04%, 0.049) 15.5%,
+    hsla(251.35, 60.66%, 88.04%, 0.104) 22.5%,
+    hsla(251.35, 60.66%, 88.04%, 0.175) 29%,
+    hsla(251.35, 60.66%, 88.04%, 0.259) 35.3%,
+    hsla(251.35, 60.66%, 88.04%, 0.352) 41.2%,
+    hsla(251.35, 60.66%, 88.04%, 0.45) 47.1%,
+    hsla(251.35, 60.66%, 88.04%, 0.55) 52.9%,
+    hsla(251.35, 60.66%, 88.04%, 0.648) 58.8%,
+    hsla(251.35, 60.66%, 88.04%, 0.741) 64.7%,
+    hsla(251.35, 60.66%, 88.04%, 0.825) 71%,
+    hsla(251.35, 60.66%, 88.04%, 0.896) 77.5%,
+    hsla(251.35, 60.66%, 88.04%, 0.951) 84.5%,
+    hsla(251.35, 60.66%, 88.04%, 0.987) 91.9%,
+    hsl(251.35, 60.66%, 88.04%) 100%
+  );
+}
+
+body.dark .convert-on-paste-overlay {
+  background: linear-gradient(
+    to bottom,
+    hsla(265.79, 79.08%, 30%, 0) 0%,
+    hsla(265.79, 79.08%, 30%, 0.013) 8.1%,
+    hsla(265.79, 79.08%, 30%, 0.049) 15.5%,
+    hsla(265.79, 79.08%, 30%, 0.104) 22.5%,
+    hsla(265.79, 79.08%, 30%, 0.175) 29%,
+    hsla(265.79, 79.08%, 30%, 0.259) 35.3%,
+    hsla(265.79, 79.08%, 30%, 0.352) 41.2%,
+    hsla(265.79, 79.08%, 30%, 0.45) 47.1%,
+    hsla(265.79, 79.08%, 30%, 0.55) 52.9%,
+    hsla(265.79, 79.08%, 30%, 0.648) 58.8%,
+    hsla(265.79, 79.08%, 30%, 0.741) 64.7%,
+    hsla(265.79, 79.08%, 30%, 0.825) 71%,
+    hsla(265.79, 79.08%, 30%, 0.896) 77.5%,
+    hsla(265.79, 79.08%, 30%, 0.951) 84.5%,
+    hsla(265.79, 79.08%, 30%, 0.987) 91.9%,
+    hsl(265.79, 79.08%, 30%) 100%
+  );
+}
