+++ app/src/pages/EditHosted.tsx
@@ -26,7 +26,7 @@ import { useTabsStore } from "../lib/useTabsStore";
 import { useHasProAccess } from "../lib/hooks";
 import { ThemeTab } from "../components/Tabs/ThemeTab";
 import { FlowchartLayout } from "../components/FlowchartLayout";
-import { AiToolbar2 } from "../components/AiToolbar2";
+import { AiToolbar } from "../components/AiToolbar";
 
 export default function EditHosted() {
   const { id } = useParams<{ id: string }>();
@@ -95,7 +95,7 @@ export default function EditHosted() {
             value="Document"
             className="bg-white dark:bg-black overflow-hidden relative grid grid-rows-[auto_minmax(0,1fr)] h-full"
           >
-            <AiToolbar2 />
+            <AiToolbar />
             <TextEditor
               value={text}
               onChange={onChange}
